tests/xstatfsx.c: fix mips o32 support
mips o32 differs from all other 32-bit architectures by defining
__kernel_fsid_t structure as an array of long ints.

* tests/xstatfsx.c (print_statfs): Explicitly cast elements
of PRINT_F_FSID array to unsigned int.

+++ tests/xstatfsx.c
@@ -81,7 +81,8 @@ print_statfs(const char *const sample, const char *magic_str)
 	PRINT_NUM(f_files);
 	PRINT_NUM(f_ffree);
 #ifdef PRINT_F_FSID
-	printf(", f_fsid={%u, %u}", b->PRINT_F_FSID[0], b->PRINT_F_FSID[1]);
+	printf(", f_fsid={%u, %u}",
+	       (unsigned) b->PRINT_F_FSID[0], (unsigned) b->PRINT_F_FSID[1]);
 #endif
 	PRINT_NUM(f_namelen);
 #ifdef PRINT_F_FRSIZE
